Handle GDTracker passing to GDDistanceMap
Can't pass the ptr from godot (not sure why exacly, it might be
a linker problem because gddistanceMap has it's own linked tracker)
So instance, Tracker has a global g_tracker and getInstance() method
to return it. GDDisstanceMap uses that method to get the Tracker instance.

However, can't use the dll built here, so dont copy it in place. The tracker.dll
build by GDDistanceMap needs used (not sure why)